багфиксы, "ускорение генератора", всякая шелуха

+++ MinesServer/GameShit/GUI/Horb/Card.cs
@@ -6,7 +6,7 @@
 
         public static Card Item(short item, string text) => new(CardImageType.Item, item.ToString(), text);
 
-        public static Card Clan(short clan, string text) => new(CardImageType.Clan, clan.ToString(), text);
+        public static Card Clan(byte clan, string text) => new(CardImageType.Clan, clan.ToString(), text);
 
         public static Card WebImage(string uri, int width, int height, string text) => new(CardImageType.WebImage, $"{width}#{height}#{uri.Replace(":", "%")}", text);
